Medithon-ThinkAI-main/requirements.txt
streamlit==1.38.0
firebase-admin==6.1.0
requests==2.32.3
python-dotenv==1.0.1

# LangChain and related libraries
langchain==0.0.205
langchain-groq==0.1.0
langchain-huggingface==0.0.1
langchain-core==0.0.4
langchain-community==0.0.10

# FAISS for vector storage
faiss-cpu==1.7.4

# Hugging Face Embeddings
sentence-transformers==2.2.2

# PyPDF for document loading
pypdf==3.15.1

faiss-cpu
pypdf
